added and improved tests

+++ src/app/Point.py
@@ -1,7 +1,7 @@
 class Point:
-    x: float
-    y: float
+    x: int
+    y: int
 
-    def __init__(self, x: float, y: float) -> None:
+    def __init__(self, x: int, y: int) -> None:
         self.x = x
         self.y = y
